changed type internals to internal 	modified:   cmd/main.go 	deleted:    internals/casting/converter.go 	deleted:    internals/config/config.go 	deleted:    internals/config/filereader.go 	deleted:    internals/elastic/indexer.go 	deleted:    internals/movedata/move.go 	deleted:    internals/scheduler/scheduler.go 	modified:   test/complete_config_test.go 	modified:   test/config_test.go 	modified:   test/insert_integration_test.go 	modified:   test/insert_typed_integration_test.go 	modified:   test/integration_common.go 	modified:   test/integration_json_test.go 	modified:   test/invalid_config_test.go 	modified:   test/scheduling_integration_test.go 	modified:   test/type_conversion_test.go 	modified:   test/upsert_integration_test.go
