+++ test/complete_config_test.go
@@ -3,7 +3,7 @@ package test
 import (
 	"testing"
 
-	"github.com/Ringloop/mr-plow/internals/config"
+	"github.com/Ringloop/mr-plow/internal/config"
 	"github.com/stretchr/testify/require"
 )
