+++ test/upsert_integration_test.go
@@ -5,8 +5,8 @@ import (
 	"log"
 	"testing"
 
-	"github.com/Ringloop/mr-plow/internals/elastic"
-	"github.com/Ringloop/mr-plow/internals/movedata"
+	"github.com/Ringloop/mr-plow/internal/elastic"
+	"github.com/Ringloop/mr-plow/internal/movedata"
 	_ "github.com/lib/pq"
 	"github.com/stretchr/testify/require"
 )
